feat: FE metadate 로고,타이틀 적용
